(Varitional) ICM-EM algorithm for implementing FAST model with structurized parameters

(Varitional) ICM-EM algorithm for implementing FAST model with structurized parameters

FAST_structure(
  XList,
  AdjList,
  q = 15,
  fit.model = c("poisson", "gaussian"),
  parameterList = NULL
)

Arguments

XList

an M-length list consisting of multiple matrices with class dgCMatrix or matrix that specify the count/log-count gene expression matrix for each data batch used for FAST model.

AdjList

an M-length list of sparse matrices with class dgCMatrix, specify the adjacency matrix used for intrisic CAR model in FAST. We provide this interface for those users who would like to define the adjacency matrix by themselves.

q

an optional integer, specify the number of low-dimensional embeddings to extract in FAST

fit.model

an optional string, specify the version of FAST to be fitted. The Gaussian version models the log-count matrices while the Poisson verions models the count matrices; default as gaussian due to fastter computation.

parameterList

an optional list, specify other parameters in FAST model; see model_set_FAST for other paramters. The default is NULL that means the default parameters produced by model_set_FAST is used.

Value

return a list including the following components: (1) hV: an M-length list consisting of spatial embeddings in FAST; (2) nu: the estimated intercept vector; (3) Psi: the estimated covariance matrix; (4) W: the estimated shared loading matrix; (5) Lam: the estimated covariance matrix of error term; (6): ELBO: the ELBO value when algorithm convergence; (7) ELBO_seq: the ELBO values for all itrations.

Details

None

References

None

See also