Fit an iSC-MEB model using specified multi-section embeddings

Integrate multiple SRT data based on the PRECASTObj by FAST and iSC-MEB model fitting.

iscmeb_run(
  VList,
  AdjList,
  K,
  beta_grid = seq(0, 5, by = 0.2),
  maxIter = 25,
  epsLogLik = 1e-05,
  verbose = TRUE,
  int.model = "EEE",
  init.start = 1,
  Sigma_equal = FALSE,
  Sigma_diag = TRUE,
  seed = 1
)

Arguments

VList

a M-length list of embeddings. The i-th element is a ni * q matrtix, where ni is the number of spots of sample i, and q is the number of embeddings. We provide this interface for those users who would like to define the embeddings by themselves.

AdjList

an M-length list of sparse matrices with class dgCMatrix, specify the adjacency matrix used for intrisic CAR model in FAST. We provide this interface for those users who would like to define the adjacency matrix by themselves.

K

an integer, specify the number of clusters.

beta_grid

an optional vector of positive value, the candidate set of the smoothing parameter to be searched by the grid-search optimization approach, defualt as a sequence starts from 0, ends with 5, increase by 0.2.

maxIter

the maximum iteration of ICM-EM algorithm. The default is 25.

epsLogLik

a string, the species, one of 'Human' and 'Mouse'.

verbose

an optional intger, spcify the number of housekeeping genes to be selected.

int.model

an optional string, specify which Gaussian mixture model is used in evaluting the initial values for iSC.MEB, default as "EEE"; and see Mclust for more models' names.

init.start

an optional number of times to calculate the initial value (1 by default). When init.start is larger than 1, initial value will be determined by log likelihood of mclust results.

Sigma_equal

an optional logical value, specify whether Sigmaks are equal, default as FALSE.

Sigma_diag

an optional logical value, specify whether Sigmaks are diagonal matrices, default as TRUE.

seed

an optional integer, the random seed in fitting iSC-MEB model.

Value

returns a iSCMEBResObj object which contains all model results.