Transfer gene names from one fortmat to the other format

Transfer gene names from one fortmat to the other format for two species: human and mouse.

transferGeneNames(
  genelist,
  now_name = "ensembl",
  to_name = "symbol",
  species = c("Human", "Mouse"),
  Method = c("eg.db", "biomart")
)

Arguments

genelist

a string vector, the gene list to be transferred.

now_name

a string, the current format of gene names, one of 'ensembl', 'symbol'.

to_name

a string, the format of gene names to transfer, one of 'ensembl', 'symbol'.

species

a string, the species, one of 'Human' and 'Mouse'.

Method

a string, the method to use, one of 'biomaRt' and 'eg.db', default as 'eg.db'.

Value

Return a string vector of transferred gene names. The gene names not matched in the database will not change.

Examples

geneNames <- c("ENSG00000171885", "ENSG00000115756")
transferGeneNames(geneNames, now_name = "ensembl", to_name="symbol",species="Human", Method='eg.db')
#> 'select()' returned 1:1 mapping between keys and columns
#> ENSG00000171885 ENSG00000115756 
#>          "AQP4"        "HPCAL1"